+++ src/common/enums/notification.event.payload.ts
@@ -19,15 +19,15 @@ export enum NotificationEventPayload {
   SPACE_COMMUNITY_INVITATION_USER_PLATFORM = 'SPACE_COMMUNITY_INVITATION_USER_PLATFORM',
   SPACE_COMMUNICATION_MESSAGE_DIRECT = 'SPACE_COMMUNICATION_MESSAGE_DIRECT',
   SPACE_COMMUNICATION_UPDATE = 'SPACE_COMMUNICATION_UPDATE',
-  SPACE_COLLABORATION_POST = 'SPACE_COLLABORATION_POST',
-  SPACE_COLLABORATION_POST_COMMENT = 'SPACE_COLLABORATION_POST_COMMENT',
-  SPACE_COLLABORATION_WHITEBOARD = 'SPACE_COLLABORATION_WHITEBOARD',
   SPACE_COLLABORATION_CALLOUT = 'SPACE_COLLABORATION_CALLOUT',
 
   // User notifications
   USER = 'USER',
   USER_MESSAGE_DIRECT = 'USER_MESSAGE_DIRECT',
   USER_MESSAGE_ROOM = 'USER_MESSAGE_ROOM',
+
+  // Virtual Contributor related payloads
+  VIRTUAL_CONTRIBUTOR = 'VIRTUAL_CONTRIBUTOR',
 }
 
 registerEnumType(NotificationEventPayload, {
