+++ src/services/adapters/notification-adapter/notification.adapter.module.ts
@@ -13,6 +13,7 @@ import { NotificationInAppAdapterModule } from '../notification-in-app-adapter/n
 import { NotificationRecipientsModule } from '@services/api/notification-recipients/notification.recipients.module';
 import { NotificationSpaceAdapter } from './notification.space.adapter';
 import { NotificationPlatformAdapter } from './notification.platform.adapter';
+import { NotificationUserAdapter } from './notification.user.adapter';
 
 @Module({
   imports: [
@@ -29,6 +30,7 @@ import { NotificationPlatformAdapter } from './notification.platform.adapter';
     NotificationAdapter,
     NotificationSpaceAdapter,
     NotificationPlatformAdapter,
+    NotificationUserAdapter,
   ],
   exports: [
     NotificationAdapter,
