+++ test/mocks/notification.space.adapter.service.mock.ts
@@ -8,8 +8,8 @@ export const MockNotificationSpaceAdapter: ValueProvider<
   provide: NotificationSpaceAdapter,
   useValue: {
     spaceCollaborationCalloutPublished: jest.fn(),
-    spaceCollaborationPostCreated: jest.fn(),
-    spaceCollaborationPostComment: jest.fn(),
+    spaceCollaborationCalloutContributionCreated: jest.fn(),
+    spaceCollaborationCalloutContributionComment: jest.fn(),
     spaceCommunicationUpdateSent: jest.fn(),
     spaceCommunityApplicationCreated: jest.fn(),
     spaceCommunityNewMember: jest.fn(),
